+++ tests/models/OrderedCategory.php
@@ -1,0 +1,13 @@
+<?php
+
+use Baum\Node;
+
+class OrderedCategory extends Node {
+
+  protected $table = 'categories';
+
+  protected $orderColumn = 'name';
+
+  public $timestamps = false;
+
+}