+++ CHANGELOG.md
@@ -14,6 +14,7 @@
   - Implement tree *index rebuilding* via `Node::rebuild`. Very useful when a tree
   structure has been messed up or to convert from a `parent_id` only table.
   - Preliminar support for soft-delete operations.
+  - Allow the user to change the default sorting column name (defaulting to `lft`).
 
 ## 1.0.9
 _Jan 14 2014_